requirements.txt
tweepy
python-decouple
pandas
nltk
matplotlib
pickle
scikit-learn